Factored out more stuff to the Task class; test_forecast passes except for CV
